+++ requirements.txt
@@ -1,0 +1,51 @@
+Package             Version
+------------------- -----------
+affine              2.4.0
+attrs               23.1.0
+certifi             2023.7.22
+cftime              1.6.2
+charset-normalizer  3.2.0
+click               8.1.6
+click-plugins       1.1.1
+cligj               0.7.2
+contourpy           1.1.0
+cycler              0.11.0
+dataretrieval       1.0.5
+Fiona               1.9.4.post1
+flopy               3.4.1
+fonttools           4.42.0
+GDAL                3.7.1
+idna                3.4
+importlib-metadata  6.8.0
+importlib-resources 6.0.1
+kiwisolver          1.4.4
+matplotlib          3.7.2
+netCDF4             1.6.4
+numpy               1.25.2
+packaging           23.1
+pandas              2.0.3
+Pillow              10.0.0
+pip                 23.2.1
+PyCRS               1.0.2
+pygsflow            1.1.0
+pyparsing           3.0.9
+pyproj              3.6.0
+pyshp               2.3.1
+python-dateutil     2.8.2
+pytz                2023.3
+rasterio            1.3.8
+rasterstats         0.19.0
+requests            2.31.0
+richdem             0.3.4
+Rtree               1.0.1
+scipy               1.11.1
+setuptools          68.0.0
+shapely             2.0.1
+simplejson          3.19.1
+six                 1.16.0
+snuggs              1.4.7
+tzdata              2023.3
+urllib3             2.0.4
+wheel               0.41.1
+xarray              2023.7.0
+zipp                3.16.2
